+++ bin/reindex_search.js
@@ -116,7 +116,7 @@ process.stdout.write(`\n`);
             }
           }
 
-          await new Promise((resolve) => setTimeout(resolve, 1000 * delay));
+          await delay(1000 * delay);
         }
       }