+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.4.0] - Not released
+### Fixed
+- User data deletion process has been optimized for likes and comment likes.
+- Job manager now starting with a random delay for more even distribution of the
+  job fetching in multi-node environment.
 
 ## [2.3.1] - 2022-07-21
 ### Fixed